Replace the `quote` function with a macro
